term-build module
