+++ src/asgard_term_manager.cpp
@@ -1,914 +1,9 @@
 #include "asgard_term_manager.hpp"
 
-#include "asgard_coefficients_mats.hpp" // also brings in small-mats module
-
 #include "asgard_blas.hpp"
 
 namespace asgard
 {
-
-template<typename P>
-term_manager<P>::term_manager(prog_opts const &options, pde_domain<P> const &domain,
-                              pde_scheme<P> &pde, sparse_grid const &grid,
-                              hierarchy_manipulator<P> const &hier,
-                              connection_patterns const &conn)
-  : num_dims(domain.num_dims()), max_level(options.max_level()), legendre(hier.degree()),
-    moms(domain, max_level, hier, std::move(pde.mlist), pde.mom_groups)
-#ifdef ASGARD_USE_MPI
-    , resources(options.mpicomm)
-#endif
-{
-  if (num_dims == 0)
-    return;
-
-  pde.finalize_term_groups(); // if using groups, else this does nothing
-
-  if (pde.mass() and not pde.mass().is_identity())
-    mass_term = std::move(pde.mass_);
-
-  std::vector<term_md<P>> &pde_terms = pde.terms_;
-  int num_terms = [&]() -> int {
-    // get the effective number of terms, chained or not
-    int n = 0;
-    for (auto const &t : pde_terms)
-      n += (t.is_chain()) ? t.num_chain() : 1;
-    return n;
-  }();
-
-  { // copy over the group ids, keep flattened format
-    term_groups.reserve(pde.term_groups.size());
-    int ibegin = 0;
-    for (auto const &tg : pde.term_groups) {
-      int n = 0;
-      for (int i : indexrange(tg))
-        n += (pde_terms[i].is_chain()) ? pde_terms[i].num_chain() : 1;
-      term_groups.emplace_back(ibegin, ibegin + n);
-      ibegin += n;
-    }
-    source_groups.resize(pde.source_groups.size());
-    for (int i : iindexof(pde.source_groups))
-      source_groups[i].source_range = pde.source_groups[i];
-  }
-
-  terms.resize(num_terms);
-
-  {
-    bool has_interp = pde.has_interp_funcs;
-
-    auto ir = terms.begin();
-    for (int i : iindexof(pde_terms.size()))
-    {
-      if (pde_terms[i].is_chain()) {
-        int const num_chain = pde_terms[i].num_chain();
-
-        // this indicates that t1 and/or t2 workspaces are needed
-        if (num_chain >= 2 and t1.empty())
-          t1.resize(1);
-        if (num_chain >= 3 and t2.empty())
-          t2.resize(1);
-
-        has_interp = has_interp or pde_terms[i].chain_[0].is_interpolatory();
-
-        *ir = term_entry<P>(std::move(pde_terms[i].chain_[0]));
-        ir++->num_chain = num_chain;
-        for (int c = 1; c < num_chain; c++) {
-          has_interp = has_interp or pde_terms[i].chain_[c].is_interpolatory();
-
-          *ir = term_entry<P>(std::move(pde_terms[i].chain_[c]));
-          ir++->mark_as_chain_link();
-        }
-      } else {
-        has_interp = has_interp or pde_terms[i].is_interpolatory();
-
-        *ir++ = term_entry<P>(std::move(pde_terms[i]));
-      }
-    }
-    if (has_interp)
-      interp = interpolation_manager<P>(options, domain, hier, conn);
-  }
-
-  int num_bc = 0;
-
-  // check if we need to keep the intermediate terms from matrix builds
-  for (auto &tt : terms) {
-    int const n = static_cast<int>(tt.tmd.bc_flux_.size());
-    tt.bc = indexrange(num_bc, num_bc + n);
-    num_bc += n;
-  }
-
-  // form groups for the boundary conditions
-  if (not term_groups.empty()) {
-    int j = 0, bc_begin = 0, bc_end = 0; // index for the boundary conditions
-    for (int groupid : iindexof(term_groups)) {
-      for (int it : indexrange(term_groups[groupid]))
-        bc_end += terms[it].bc.size();
-      source_groups[j++].bc_range = irange(bc_begin, bc_end);
-      bc_begin = bc_end;
-    }
-  }
-
-  bcs.reserve(num_bc);
-  for (int tid : iindexof(terms)) {
-    term_entry<P> &tt = terms[tid];
-    int const fdim = tt.tmd.flux_dim();
-    tt.flux_dim = fdim;
-    for (auto &b : tt.tmd.bc_flux_) {
-      bcs.emplace_back(std::move(b));
-      bcs.back().term_index = tid;
-      for (int d : iindexof(num_dims)) {
-        if (bcs.back().flux.chain_level(d) == -1) { // reset to the lowest level
-          bcs.back().flux.chain_level(d) = (tt.tmd.dim(d).is_chain())
-                                          ? (tt.tmd.dim(d).num_chain() - 1) : 0;
-        }
-      }
-      if (bcs.back().flux.func().ignores_time()) {
-        bcs.back().tmode = boundary_entry<P>::time_mode::constant;
-      } else {
-        if (bcs.back().flux.func_.ftime()) {
-          if (bcs.back().flux.func_.cdomain(fdim) == 0) {
-            bcs_have_time_dep = true;
-            bcs.back().tmode  = boundary_entry<P>::time_mode::time_dependent;
-            for (int d : iindexof(num_dims)) {
-              rassert(not tt.tmd.dim(d).is_chain(),
-                      "cannot use non-separable in time boundary conditions with 1d-chains, "
-                      "the purpose of the 1d chain is to pre-compute and cache entries but non-separable "
-                      "data cannot be pre-computed, an md-chain must be used instead");
-            }
-          } else {
-            bcs.back().tmode = boundary_entry<P>::time_mode::separable;
-          }
-        } else {
-          // fdim is constant, but the other dirs are non-separable
-          bcs.back().tmode = boundary_entry<P>::time_mode::separable;
-        }
-      }
-    }
-  }
-
-  // domain left/right bounds
-  for (int d : iindexof(num_dims)) {
-    xleft[d]  = pde.domain().xleft(d);
-    xright[d] = pde.domain().xright(d);
-  }
-
-  build_mass_matrices(hier, conn); // large, up to max-level
-  rebuild_mass_matrices(grid); // small, up to the current level
-
-  std::vector<separable_func<P>> &sep = pde.sources_sep_;
-
-  int num_sources = 0;
-  for (auto const &s : sep) {
-    int const dims = s.num_dims();
-    rassert(dims == 0 or dims == num_dims, "incorrect dimension set for source");
-    if (dims > 0) ++num_sources;
-  }
-
-  sources_md = std::move(pde.sources_md_);
-  sources.reserve(num_sources);
-
-  for (auto &s : sep) {
-    if (s.num_dims() == 0)
-      continue;
-
-    if (s.ignores_time() or s.ftime()) {
-      // using constant entry
-      if (s.ignores_time()) {
-        sources.emplace_back(source_entry<P>::time_mode::constant);
-        sources.back().func = 0; // no need for a func
-      } else {
-        sources.emplace_back(source_entry<P>::time_mode::separable);
-        sources.back().func = s.ftime();
-      }
-
-      for (int d : iindexof(num_dims)) {
-        if (s.is_const(d)) {
-          sources.back().consts[d]
-              = hier.get_project1d_c(s.cdomain(d), mass[d], d, max_level);
-        } else {
-          sources.back().consts[d] = hier.get_project1d_f(
-              [&](std::vector<P> const &x, std::vector<P> &y)-> void { s.fdomain(d, x, 0, y); },
-              mass[d], d, max_level);
-        }
-      }
-
-    } else {
-      // non-separable in time
-      sources_have_time_dep = true;
-      sources.emplace_back(source_entry<P>::time_mode::time_dependent);
-      sources.back().func = std::move(s);
-    }
-  }
-
-  prapare_kron_workspace(grid); // setup kronmult workspace
-
-  has_terms_ = not terms.empty();
-  assign_compute_resources();
-
-  // prepare the workspaces for the sources
-  // consider only sources that are associated with this MPI rank and not time-dependant
-  // the time sources cannot use workspace to accelerate computations
-  #ifdef ASGARD_USE_MPI
-  auto is_active_src = [&, this](source_entry<P> const &src) -> bool
-    {
-      if (not resources.owns(src.rec))
-        return false;
-      return (not src.is_time_dependent());
-    };
-  auto is_active_bc = [&, this](boundary_entry<P> const &bc) -> bool
-    {
-      if (not resources.owns(terms[bc.term_index].rec))
-        return false;
-      return (not bc.is_time_dependent());
-    };
-  #else
-  auto is_active_src = [&](source_entry<P> const &src) -> bool
-    {
-      return (not src.is_time_dependent());
-    };
-  auto is_active_bc = [&](boundary_entry<P> const &bc) -> bool
-    {
-      return (not bc.is_time_dependent());
-    };
-  #endif
-
-  for (auto const &src : sources)
-    if (is_active_src(src)) num_lumped++;
-
-  for (auto const &bc : bcs)
-    if (is_active_bc(bc)) num_lumped++;
-
-  if (not source_groups.empty()) { // set sources group by group
-    int ibegin = 0, iend = 0;
-    for (size_t i = 0; i < source_groups.size(); i++) {
-      for (int is : indexrange(source_groups[i].source_range))
-        if (is_active_src(sources[is]))
-          sources[is].ilump = iend++;
-      for (int ib : indexrange(source_groups[i].bc_range))
-        if (is_active_bc(bcs[ib]))
-          bcs[ib].ilump = iend++;
-      source_groups[i].lump_range = irange(ibegin, iend);
-      ibegin = iend;
-    }
-  } else { // no groups, lump everything together
-    int j = 0;
-    for (auto &src : sources)
-      if (is_active_src(src)) src.ilump = j++;
-    for (auto &bc : bcs)
-      if (is_active_bc(bc)) bc.ilump = j++;
-  }
-  sweights.reserve(num_lumped); // one weight per lumped source
-
-  // second pass on the problem of assigning workspaces and preparing objects
-  // e.g., the needed resources change if this MPI rank has no terms with need
-  {
-    // set interpolatory properties
-    for (int i : indexof(terms)) {
-      auto &t = terms[i];
-      t.is_interpolatory = t.tmd.is_interpolatory();
-      if (t.is_interpolatory) {
-        if (t.num_chain == 1) { // single entry
-          t.interp_uses_ifield = true;
-        } else if (t.is_chain_link() and
-                   (i+1 == static_cast<int>(terms.size())
-                    or not terms[i+1].is_chain_link())) {
-          // if part of a chain and the next term is not from the current chain
-          // i.e., this is the first link in the chain
-          t.interp_uses_ifield = true;
-        }
-      }
-    }
-
-    bool has_field_interp = false; // interpolating from a field
-    auto it = terms.begin();
-    while (it < terms.end())
-    {
-      #ifdef ASGARD_USE_MPI
-      if (not resources.owns(it->rec)) {
-        it += it->num_chain;
-        continue;
-      }
-      #endif
-      has_field_interp = has_field_interp or it->interp_uses_ifield;
-      if (it->is_chain_start())
-        has_field_interp = has_field_interp or (it + it->num_chain -1)->interp_uses_ifield;
-      it += it->num_chain;
-    }
-
-    if (has_field_interp)
-      ifield.resize(1);
-
-    // handle the moment dependence
-    std::vector<moment_id> active_moments;
-    active_moments.reserve(250); // should be more than enough, not a big deal otherwise
-    bool has_poisson = false;
-    for (auto const &tentry : terms) {
-      #ifdef ASGARD_USE_MPI
-      if (not resources.owns(tentry.rec))
-        continue;
-      #endif
-      has_poisson = has_poisson or tentry.has_poisson;
-      if (tentry.is_separable()) // only separable terms can have 1D moment deps
-        for (int d : iindexof(num_dims)) {
-          auto const &mids = tentry.tmd.dim(d).mids_;
-          if (not mids.empty()) {
-            active_moments.insert(active_moments.end(), mids.begin(), mids.end());
-          }
-        }
-    }
-    if (has_poisson) {
-      if (term_groups.empty())
-        has_poisson_.resize(1, true);
-      else
-        has_poisson_.resize(term_groups.size(), false); // will process groups below
-    }
-    if (not term_groups.empty()) {
-      for (int gid : iindexof(term_groups)) {
-        bool needs = false;
-        for (int tid : indexrange(term_groups[gid])) {
-          #ifdef ASGARD_USE_MPI
-          if (not resources.owns(terms[tid].rec))
-            continue;
-          #endif
-          needs = needs or terms[tid].has_poisson;
-        }
-        if (needs)
-          has_poisson_[gid] = needs;
-      }
-    }
-  }
-}
-
-template<typename P>
-void term_manager<P>::buld_term(
-    int const tid, sparse_grid const &grid, connection_patterns const &conn,
-    hierarchy_manipulator<P> const &hier, precon_method precon, P alpha)
-{
-  if (terms[tid].tmd.is_interpolatory()) // skip interpolation terms
-    return;
-
-  expect(legendre.pdof == hier.degree() + 1);
-  expect(not terms[tid].tmd.is_chain());
-
-  auto &tmd = terms[tid];
-
-  bool merging_with_interp = false;
-  if ((tmd.is_chain_start() or tmd.is_chain_link())
-       and (static_cast<size_t>(tid + 1) < terms.size())
-        and terms[tid + 1].is_chain_link()
-         and terms[tid + 1].is_interpolatory)
-  {
-    // there is a potential here to merge this separable term with hier2wav
-    merging_with_interp = true;
-    for (int d : iindexof(num_dims))
-      if (tmd.tmd.dim(d).change() != changes_with::none)
-        merging_with_interp = false;
-    // if the 1d terms are changing, then skip the merge
-    // if everything is constant, we can merge
-  }
-
-  if (merging_with_interp)
-  {
-    constexpr bool merge_with_interp = true;
-    terms[tid + 1].interp_stop_at_hierarchy = true;
-
-    std::vector<int> id_dirs;
-    id_dirs.reserve(num_dims);
-    for (int d : iindexof(num_dims))
-    {
-      rebuld_term1d(terms[tid], d, max_level, conn, hier, precon, alpha, merge_with_interp);
-      if (terms[tid].tmd.dim(d).is_identity())
-        id_dirs.push_back(d);
-    }
-    // adjust the kronmult permutations using the fact that the identity directions
-    // were replaced by the hier2wav matrix, which is upper hierarchical
-    if (not id_dirs.empty())
-      terms[tid].perm.prepad_upper(id_dirs);
-  }
-  else
-  {
-    for (int d : iindexof(num_dims)) {
-      auto const &t1d = tmd.tmd.dim(d);
-
-      int level = grid.current_level(d); // required level
-
-      // terms that don't change should be build only once
-      if (t1d.change() == changes_with::none)
-        level = max_level; // build up to the max
-
-      rebuld_term1d(terms[tid], d, level, conn, hier, precon, alpha);
-    } // move to next dimension d
-  }
-}
-
-template<typename P>
-void term_manager<P>::rebuld_term1d(
-    term_entry<P> &tentry, int const dim, int level,
-    connection_patterns const &conn, hierarchy_manipulator<P> const &hier,
-    precon_method precon, P alpha, bool merge_with_interp)
-{
-  int const n = hier.degree() + 1;
-  auto &t1d   = tentry.tmd.dim(dim);
-
-  block_diag_matrix<P> *bmass = nullptr; // mass to use for the boundary source
-
-  // apply the mass matrix, if any
-  if (tentry.num_chain < 0) {
-    // member of a chain, can have unique mass matrix
-    mass_md<P> const &tms = tentry.tmd.mass();
-    if (tms and not tms[dim].is_identity()) {
-      int const nrows = fm::ipow2(level); // needed number of rows
-      if (tentry.mass[dim].nrows() != nrows) {
-        build_raw_mass(dim, tms[dim], max_level, tentry.mass[dim]);
-        tentry.mass[dim].spd_factorize(n);
-      }
-      bmass = &tentry.mass[dim];
-    }
-  } else if (mass[dim]) { // no chain (or last link), and there's global mass
-    // global case, use the global mass matrices
-    if (level == max_level) {
-      bmass = &mass[dim];
-    } else { // using lower level, construct lower mass matrix
-      int const nrows = fm::ipow2(level); // needed number of rows
-      if (lmass[dim].nrows() != nrows) {
-        build_raw_mass(dim, mass_term[dim], max_level, lmass[dim]);
-        lmass[dim].spd_factorize(n);
-      }
-      bmass = &lmass[dim];
-    }
-  }
-
-  bool is_diag = t1d.is_volume();
-  if (t1d.is_chain()) {
-    rebuld_chain(tentry, dim, level, hier, bmass, is_diag, wraw_diag, wraw_tri);
-  } else {
-    build_raw_mat(tentry, dim, 0, level, hier, bmass, wraw_diag, wraw_tri);
-  }
-
-  // the build/rebuild put the result in raw_diag or raw_tri
-  // if the term is identity, then there is no matrix, all the calls
-  // above are needed to handle the boundary conditions
-  if (t1d.is_identity()) {
-    if (merge_with_interp)
-      tentry.coeffs[dim] = interp.get_hier2wav();
-  } else {
-    if (is_diag) {
-      if (merge_with_interp)
-        tentry.coeffs[dim] = interp.mult_transform_h2w(hier, conn, wraw_diag, raw_diag0);
-      else
-        tentry.coeffs[dim] = hier.diag2hierarchical(wraw_diag, level, conn);
-    } else {
-      if (merge_with_interp)
-        tentry.coeffs[dim] = interp.mult_transform_h2w(hier, conn, wraw_tri, raw_tri0);
-      else
-        tentry.coeffs[dim] = hier.tri2hierarchical(wraw_tri, level, conn);
-    }
-  }
-
-  // the last interpolation stage (2wav) comes with a scaling factor
-  // apply the scaling factor to the zeroth dimension
-  if (merge_with_interp and dim == 0)
-    tentry.coeffs[dim].scal(interp.wav_scale_h2w());
-
-  #ifdef ASGARD_USE_GPU
-  if (not tentry.coeffs[dim].empty()) { // load to the GPU
-    compute->set_device(gpu::device{tentry.rec.device});
-    tentry.gpu_lcoeffs[dim].resize(level + 1);
-    std::vector<P*> coeff_pntrs(level + 1, nullptr);
-    for (int l = 0; l < level; l++) {
-      tentry.gpu_lcoeffs[dim][l] = tentry.coeffs[dim].get_subpattern(l, conn).data_vector();
-      coeff_pntrs[l] = tentry.gpu_lcoeffs[dim][l].data();
-    }
-    tentry.gpu_lcoeffs[dim][level] = tentry.coeffs[dim].data_vector();
-    coeff_pntrs[level]             = tentry.gpu_lcoeffs[dim][level].data();
-
-    tentry.gpu_coeffs[dim] = coeff_pntrs;
-
-    compute->set_device(gpu::device{0});
-  }
-  #endif
-
-  // apply the mass matrices and convert to hierarchical form
-  for (int b : indexrange{tentry.bc}) {
-    boundary_entry<P> &bentry = bcs[b];
-    if (not bentry.consts[dim].empty()) {
-      // will be empty if non-flux direction and non-separable in time
-      hier.transform(level, bentry.consts[dim]);
-    }
-  }
-
-  // build the ADI preconditioner here
-  if (precon == precon_method::adi) {
-    if (is_diag) {
-      to_euler(legendre.pdof, alpha, wraw_diag);
-      psedoinvert(legendre.pdof, wraw_diag, raw_diag0);
-      tentry.adi[dim] = hier.diag2hierarchical(raw_diag0, level, conn);
-    } else {
-      to_euler(legendre.pdof, alpha, wraw_tri);
-      psedoinvert(legendre.pdof, wraw_tri, raw_tri0);
-      tentry.adi[dim] = hier.tri2hierarchical(raw_tri0, level, conn);
-    }
-  }
-}
-
-template<typename P>
-void term_manager<P>::build_raw_mat(
-    term_entry<P> &tentry, int d, int clink, int level,
-    hierarchy_manipulator<P> const &hier,
-    block_diag_matrix<P> const *bmass,
-    block_diag_matrix<P> &raw_diag, block_tri_matrix<P> &raw_tri)
-{
-  term_1d<P> &t1d = (tentry.tmd.dim(d).is_chain()) ? tentry.tmd.dim(d).chain_[clink] : tentry.tmd.dim(d);
-  expect(not t1d.is_chain());
-
-  switch (t1d.optype())
-  {
-    case operation_type::volume:
-      switch (t1d.depends()) {
-        case term_dependence::electric_field_only:
-          if (t1d.rhs()) {
-            // using w1 as workspaces, it probably has enough space already
-            size_t const n = kwork.w1.size();
-            t1d.rhs(moms.poisson_level(), kwork.w1);
-            gen_diag_cmat_pwc<P>(legendre, level, kwork.w1, raw_diag);
-            kwork.w1.resize(n);
-          } else {
-            gen_diag_cmat_pwc<P>(legendre, level, moms.poisson_level(), raw_diag);
-          }
-          break;
-        case term_dependence::electric_field:
-          throw std::runtime_error("el-field with position depend is not done (yet)");
-          break;
-        case term_dependence::moment_divided_by_density:
-          gen_diag_mom_over_zero<P>(legendre, level, t1d.rhs_const(),
-                                    moms.get_cached_level(t1d.moment_ids()[0], hier),
-                                    moms.get_cached_level(t1d.moment_ids()[1], hier),
-                                    raw_diag);
-          break;
-        case term_dependence::lenard_bernstein_coll_theta:
-          switch (moms.num_vel()) {
-          case 1:
-            moms.cache_levels(3, hier, t1d.mids_);
-            gen_diag_lenard_bernstein_theta<P, 1>(legendre, level, t1d.rhs_const(),
-                                                  t1d.mids_, moms.get_cached_levels(),
-                                                  raw_diag);
-            break;
-          case 2:
-            moms.cache_levels(5, hier, t1d.mids_);
-            gen_diag_lenard_bernstein_theta<P, 2>(legendre, level, t1d.rhs_const(),
-                                                  t1d.mids_, moms.get_cached_levels(),
-                                                  raw_diag);
-            break;
-          case 3:
-            moms.cache_levels(7, hier, t1d.mids_);
-            gen_diag_lenard_bernstein_theta<P, 3>(legendre, level, t1d.rhs_const(),
-                                                  t1d.mids_, moms.get_cached_levels(),
-                                                  raw_diag);
-            break;
-          default:
-            // unreachable here
-            break;
-          };
-          break;
-        default:
-          if (t1d.rhs()) {
-            gen_diag_cmat<P, operation_type::volume>
-              (legendre, xleft[d], xright[d], level, t1d.rhs(), raw_rhs, raw_diag);
-          } else {
-            gen_diag_cmat<P, operation_type::volume>
-              (legendre, level, t1d.rhs_const(), raw_diag);
-          }
-          break;
-      }
-      break;
-    case operation_type::div:
-      if (t1d.rhs()) {
-        gen_tri_cmat<P, operation_type::div, rhs_type::is_func>
-          (legendre, xleft[d], xright[d], level, t1d.rhs(), 0, t1d.flux(), t1d.boundary(), raw_rhs, raw_tri);
-      } else {
-        gen_tri_cmat<P, operation_type::div, rhs_type::is_const>
-          (legendre, xleft[d], xright[d], level, nullptr, t1d.rhs_const(), t1d.flux(), t1d.boundary(), raw_rhs, raw_tri);
-      }
-      if (t1d.penalty() != 0) {
-        gen_tri_cmat<P, operation_type::penalty, rhs_type::is_const, data_mode::increment>
-          (legendre, xleft[d], xright[d], level, nullptr, t1d.penalty(), t1d.flux(), t1d.boundary(), raw_rhs, raw_tri);
-      }
-      break;
-    case operation_type::grad:
-      if (t1d.rhs()) {
-        gen_tri_cmat<P, operation_type::grad, rhs_type::is_func>
-          (legendre, xleft[d], xright[d], level, t1d.rhs(), 0, t1d.flux(), t1d.boundary(), raw_rhs, raw_tri);
-      } else {
-        gen_tri_cmat<P, operation_type::grad, rhs_type::is_const>
-          (legendre, xleft[d], xright[d], level, nullptr, t1d.rhs_const(), t1d.flux(), t1d.boundary(), raw_rhs, raw_tri);
-      }
-      if (t1d.penalty() != 0) {
-        gen_tri_cmat<P, operation_type::penalty, rhs_type::is_const, data_mode::increment>
-          (legendre, xleft[d], xright[d], level, nullptr, t1d.penalty(), t1d.flux(), t1d.boundary(), raw_rhs, raw_tri);
-      }
-      break;
-    case operation_type::penalty:
-      expect(not t1d.rhs());
-      gen_tri_cmat<P, operation_type::penalty, rhs_type::is_const>
-        (legendre, xleft[d], xright[d], level, nullptr, t1d.rhs_const(), t1d.flux(), t1d.boundary(), raw_rhs, raw_tri);
-      break;
-    default: // case operation_type::identity:
-      // identity, nothing to do for the matrix, but may have to do boundary conditions
-      break;
-  }
-
-  if (bmass) {
-    if (t1d.optype() == operation_type::volume)
-      bmass->solve(legendre.pdof, raw_diag);
-    else
-      bmass->solve(legendre.pdof, raw_tri);
-  }
-
-  for (int b : indexrange(tentry.bc)) {
-    // handle the non-separable in time, keep rhs values
-    boundary_entry<P> &bentry = bcs[b];
-
-    if (bentry.flux.chain_level(d) > clink) {
-      expect(not bentry.consts[d].empty());
-      if (t1d.is_volume()) {
-        raw_diag.inplace_gemv(legendre.pdof, bentry.consts[d], t1);
-      } else {
-        raw_tri.inplace_gemv(legendre.pdof, bentry.consts[d], t1);
-      }
-    } else if (bentry.flux.chain_level(d) == clink) {
-      // create a new entry
-      if (tentry.flux_dim == d) {
-        int const pdof = legendre.pdof;
-
-        int64_t const num_cells = fm::ipow2(level);
-        int64_t const num_entries = pdof * num_cells;
-
-        bentry.consts[d].resize(num_entries);
-
-        P scale = P{1} / std::sqrt( (xright[d] - xleft[d]) / num_cells );
-        if (t1d.is_penalty()) // penalty flips the sign of the boundary conditions
-          scale = -scale;
-
-        if (bentry.flux.is_left()) {
-          P rhs_left  = (t1d.rhs()) ? raw_rhs.vals.front() : t1d.rhs_const();
-          if (t1d.penalty() != 0)
-            rhs_left *= P{1} + t1d.penalty();
-
-          P const fc = bentry.flux.func().cdomain(d);
-          if (fc == 0) { // non-separable in time
-            // single-point value is always separable, so we can pre-compute in d-direction
-            smmat::axpy(pdof, - rhs_left * scale, legendre.leg_left, bentry.consts[d].data());
-          } else {
-            smmat::axpy(pdof, - rhs_left * scale * fc, legendre.leg_left, bentry.consts[d].data());
-          }
-        }
-
-        if (bentry.flux.is_right()) {
-          P rhs_right = (t1d.rhs()) ? raw_rhs.vals.back()  : t1d.rhs_const();
-
-          if (t1d.penalty() != 0)
-            rhs_right *= P{1} - t1d.penalty();
-
-          P const fc = bentry.flux.func().cdomain(d);
-          if (fc == 0) { // non-separable in time
-            // single-point value is always separable, so we can pre-compute in d-direction
-            smmat::axpy(pdof, rhs_right * scale, legendre.leg_right,
-                        bentry.consts[d].data() + num_entries - pdof);
-          } else {
-            smmat::axpy(pdof, rhs_right * scale * fc, legendre.leg_right,
-                        bentry.consts[d].data() + num_entries - pdof);
-          }
-        }
-
-        if (bmass)
-          bmass->solve(pdof, bentry.consts[d]);
-
-      } else {
-        if (bentry.is_time_dependent()) // no constant components to pre-compute
-          continue;
-
-        P const dsqr = std::sqrt(xright[d] - xleft[d]);
-
-        if (bentry.flux.func().is_const(d)) {
-          if (t1d.rhs()) { // constant times spatially variable
-            bentry.consts[d] = legendre.project(t1d.is_volume(), level, dsqr,
-                                                bentry.flux.func().cdomain(d), raw_rhs.vals);
-          } else { // constant times a constant
-            P const rconst = (t1d.is_identity()) ? 1 : t1d.rhs_const();
-            bentry.consts[d] = legendre.project(level, dsqr,
-                                                bentry.flux.func().cdomain(d) * rconst);
-          }
-        } else {
-          if (t1d.rhs()) { // product of non-consts
-            std::vector<P> f(raw_rhs.pnts.size());
-            bentry.flux.func().fdomain(d, raw_rhs.pnts, 0, f);
-            bentry.consts[d] = legendre.project(t1d.is_volume(), level, dsqr, f, raw_rhs.vals);
-          } else {
-            // need function values, rhs is a constant
-            legendre.interior_quad(xleft[d], xright[d], level, raw_rhs.pnts);
-            raw_rhs.vals.resize(raw_rhs.pnts.size());
-            bentry.flux.func().fdomain(d, raw_rhs.pnts, 0, raw_rhs.vals);
-            bool constexpr use_interior = true;
-            bentry.consts[d] = legendre.project(use_interior, level, dsqr, t1d.rhs_const(), raw_rhs.vals);
-          }
-        }
-
-        if (bmass)
-          bmass->solve(legendre.pdof, bentry.consts[d]);
-      }
-    } // if the bentry is associated with a higher link, then do nothing here
-  }
-}
-
-template<typename P>
-void term_manager<P>::build_raw_mass(int dim, term_1d<P> const &t1d, int level,
-                                     block_diag_matrix<P> &raw_diag)
-{
-  expect(t1d.is_volume());
-  expect(t1d.depends() == term_dependence::none);
-
-  if (t1d.rhs()) {
-    gen_diag_cmat<P, operation_type::volume>
-      (legendre, xleft[dim], xright[dim], level, t1d.rhs(), raw_rhs, raw_diag);
-  } else {
-    gen_diag_cmat<P, operation_type::volume>
-      (legendre, level, t1d.rhs_const(), raw_diag);
-  }
-}
-
-template<typename P>
-void term_manager<P>::rebuld_chain(
-    term_entry<P> &tentry, int const d, int const level,
-    hierarchy_manipulator<P> const &hier,
-    block_diag_matrix<P> const *bmass,
-    bool &is_diag, block_diag_matrix<P> &raw_diag, block_tri_matrix<P> &raw_tri)
-{
-  term_1d<P> &t1d = tentry.tmd.dim(d);
-  expect(t1d.is_chain());
-  int const num_chain = t1d.num_chain();
-  expect(num_chain > 1);
-
-  is_diag = true;
-  for (int i : iindexof(num_chain)) {
-    if (not t1d[i].is_volume()) {
-      is_diag = false;
-      break;
-    }
-  }
-
-  if (is_diag) { // a bunch of diag matrices, easy case
-    // raw_tri will not be referenced, it's just passed in
-    // using raw_diag to make the intermediate matrices, until the last one
-    // the last product has to be written to raw_diag
-    block_diag_matrix<P> *diag0 = &raw_diag0;
-    block_diag_matrix<P> *diag1 = &raw_diag1;
-    build_raw_mat(tentry, d, num_chain - 1, level, hier, bmass, *diag0, raw_tri);
-    for (int i = num_chain - 2; i > 0; i--) {
-      build_raw_mat(tentry, d, i, level, hier, bmass, raw_diag, raw_tri);
-      diag1->check_resize(raw_diag);
-      gemm_block_diag(legendre.pdof, raw_diag, *diag0, *diag1);
-      std::swap(diag0, diag1);
-    }
-    build_raw_mat(tentry, d, 0, level, hier, bmass, *diag1, raw_tri);
-    raw_diag.check_resize(*diag1);
-    gemm_block_diag(legendre.pdof, *diag1, *diag0, raw_diag);
-
-    return;
-  }
-
-  // the final is always a tri-diagonal matrix
-  // but we have to keep track of upper/lower and diagonal
-  block_diag_matrix<P> *diag0 = &raw_diag0;
-  block_diag_matrix<P> *diag1 = &raw_diag1;
-
-  block_tri_matrix<P> *tri0 = &raw_tri0;
-  block_tri_matrix<P> *tri1 = &raw_tri1;
-
-  enum class fill {
-    diag, tri
-  };
-
-  // here we start with either a diagonal or tri-diagonal matrix
-  // and at each stage we multiply by diag/tri-matrix
-  // if we start with a diagonal, we will switch to tri at some point
-
-  fill current = (t1d.chain_.back().is_volume()) ? fill::diag : fill::tri;
-  build_raw_mat(tentry, d, num_chain - 1, level, hier, bmass, *diag0, *tri0);
-
-  for (int i = num_chain - 2; i > 0; i--)
-  {
-    build_raw_mat(tentry, d, i, level, hier, bmass, raw_diag, raw_tri);
-    // the result is in either raw_diag or raw_tri and must be multiplied and put
-    // into either diag1 or tri1, then those should swap with diag0 and tri0
-    if (t1d[i].is_volume()) { // computed a diagonal fill
-      if (current == fill::diag) { // diag-to-diag
-        diag1->check_resize(raw_diag);
-        gemm_block_diag(legendre.pdof, raw_diag, *diag0, *diag1);
-        std::swap(diag0, diag1);
-      } else { // multiplying diag by tri-diag
-        tri1->check_resize(raw_diag);
-        gemm_diag_tri(legendre.pdof, raw_diag, *tri0, *tri1);
-        std::swap(tri0, tri1);
-      }
-    } else { // computed tri matrix (upper or lower diagonal)
-      if (current == fill::diag ) { // tri times diag
-        tri1->check_resize(raw_tri);
-        gemm_tri_diag(legendre.pdof, raw_tri, *diag0, *tri1);
-        std::swap(tri0, tri1);
-        current = fill::tri;
-      } else {
-        tri1->check_resize(raw_tri);
-        gemm_block_tri(legendre.pdof, raw_tri, *tri0, *tri1);
-        std::swap(tri0, tri1);
-        current = fill::tri;
-      }
-    }
-  }
-
-  // last term, compute in diag1/tri1 and multiply into raw_tri
-  build_raw_mat(tentry, d, 0, level, hier, bmass, *diag1, *tri1);
-
-  if (t1d[0].is_volume()) {
-    // the rest must be a tri-diagonal matrix already
-    // otherwise the whole chain would consist of only diagonal ones
-    raw_tri.check_resize(*tri0);
-    gemm_diag_tri(legendre.pdof, *diag1, *tri0, raw_tri);
-  } else {
-    if (current == fill::diag) {
-      raw_tri.check_resize(*tri1);
-      gemm_tri_diag(legendre.pdof, *tri1, *diag0, raw_tri);
-    } else {
-      raw_tri.check_resize(*tri1);
-      gemm_block_tri(legendre.pdof, *tri1, *tri0, raw_tri);
-    }
-  }
-
-  // apply the penalty that is added to the whole chain
-  if (t1d.penalty() == 0)
-    return;
-
-  if (bmass) {
-    gen_tri_cmat<P, operation_type::penalty, rhs_type::is_const>
-      (legendre, xleft[d], xright[d], level, nullptr, t1d.penalty(), t1d.chain_.back().flux(),
-        t1d.chain_.back().boundary(), raw_rhs, *tri0);
-    bmass->solve(legendre.pdof, *tri0);
-    raw_tri += *tri0;
-  } else {
-    // no need to worry about the mass, just add the penalty to the raw-tri
-    gen_tri_cmat<P, operation_type::penalty, rhs_type::is_const, data_mode::increment>
-      (legendre, xleft[d], xright[d], level, nullptr, t1d.penalty(), t1d.chain_.back().flux(),
-        t1d.chain_.back().boundary(), raw_rhs, raw_tri);
-  }
-  // handle the penalty component of the boundary conditions
-  std::vector<P> penwork; // extra allocation, should be rare, when having mass + builtin penalty
-  for (int b : indexrange(tentry.bc)) {
-    // handle the non-separable in time, keep rhs values
-    boundary_entry<P> &bentry = bcs[b];
-
-    // apply only the conditions for the bottom link
-    if (bentry.flux.chain_level(d) != num_chain - 1)
-      continue;
-
-    int const pdof = legendre.pdof;
-
-    int64_t const num_cells = fm::ipow2(level);
-    int64_t const num_entries = pdof * num_cells;
-
-    if (bmass)
-      penwork.reserve(num_entries);
-
-    // for no mass, write directly into consts, else must use scratch space to invert the matrix
-    P *dest = (bmass) ? penwork.data() : bentry.consts[d].data();
-
-    expect(bentry.consts[d].size() == static_cast<size_t>(num_entries));
-
-    P const scale = -t1d.penalty() / std::sqrt( (xright[d] - xleft[d]) / num_cells );
-
-    if (bentry.flux.is_left()) {
-      P const fc = bentry.flux.func().cdomain(d);
-      if (fc == 0) { // non-separable in time
-        smmat::axpy(pdof, -scale, legendre.leg_left, dest);
-      } else {
-        smmat::axpy(pdof, -scale * fc, legendre.leg_left, dest);
-      }
-    }
-
-    if (bentry.flux.is_right()) {
-      P const fc = bentry.flux.func().cdomain(d);
-      if (fc == 0) { // non-separable in time
-        smmat::axpy(pdof, scale, legendre.leg_right, dest + num_entries - pdof);
-      } else {
-        smmat::axpy(pdof, scale * fc, legendre.leg_right, dest + num_entries - pdof);
-      }
-    }
-
-    if (bmass) {
-      bmass->solve(pdof, dest);
-      ASGARD_OMP_PARFOR_SIMD
-      for (int64_t i = 0; i < num_entries; i++)
-        bentry.consts[d][i] += dest[i];
-    }
-  }
-}
 
 template<typename P>
 void term_manager<P>::mass_apply(
@@ -1657,7 +752,6 @@ void term_manager<P>::assign_compute_resources()
 }
 
 #ifdef ASGARD_ENABLE_DOUBLE
-template struct term_entry<double>;
 template struct term_manager<double>;
 
 template void term_manager<double>::kron_diag<data_mode::increment>(
@@ -1692,7 +786,6 @@ template void term_manager<double>::apply_tmpl_gpu<double const[], double[], com
 #endif
 
 #ifdef ASGARD_ENABLE_FLOAT
-template struct term_entry<float>;
 template struct term_manager<float>;
 
 template void term_manager<float>::kron_diag<data_mode::increment>(
